pyproject.toml
[project]
name = "quiz-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Gabz047", email = "gabe1612mail@gmail.com"},
]
dependencies = [
    "django>=5.0.7",
    "djangorestframework>=3.15.2",
    "setuptools>=70.3.0",
    "django-cors-headers>=4.4.0",
    "djangorestframework-simplejwt>=5.3.1",
    "asgiref==3.8.1",
    "django==5.0.7",
    "django-cors-headers==4.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "pyjwt==2.8.0",
    "setuptools==70.3.0",
    "sqlparse==0.5.0",
    "typing-extensions==4.12.2; python_version < \"3.11\"",
    "tzdata==2024.1; sys_platform == \"win32\"",
    "drf-spectacular>=0.27.2",
    "django-extensions>=3.2.3",
    "drf-passage-identity>=0.0.3",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python manage.py runserver localhost:8000"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

